requirements.txt
streamlit>=1.38.0
pandas>=2.1.0
numpy>=1.24.0
scikit-learn>=1.4.0
supabase>=2.0.0
python-dotenv>=1.0.0
google-api-python-client>=2.100.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0
gspread>=6.0.0
pdfplumber>=0.10.0
joblib>=1.3.0
plotly>=5.18.0
httpx>=0.25.0
